I am stumped. Function cartItemTemplate seems to be the issue.
