+++ src/checkout/index.html
@@ -9,6 +9,7 @@
     <title>Sleep Outside | Checkout</title>
 
     <link rel="stylesheet" href="css/style.css" />
+    <script src="/js/checkout.js"></script>
   </head>
 
   <body>
