+++ src/js/main.js
@@ -1,5 +1,5 @@
-//import productList from "./productList.mjs";
+import productList from "./productList.mjs";
 import { loadHeaderFooter } from "./utils.mjs";
 
 loadHeaderFooter();
-//productList(".product-list", "tents");
+productList(".product-list", "tents");
